begin setting up smoke tests and reorganize styling files
