+++ src/components/NavBar.test.js
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import NavBar from './App';
 import {shallow} from "enzyme";
 
-describe("App", () => {
+describe("NavBar", () => {
   it("Renders without crashing", () => {
-    shallow(<App />)
+    shallow(<NavBar />)
   })
 })
