+++ src/components/NavBarMobile.js
@@ -3,7 +3,7 @@ import { Link } from "react-router-dom";
 import styles from "../componentStyles/NavBarMobile.css"
 import NavDrawer from "./NavDrawer"
 import {connect} from "react-redux"
-import {expandMenu} from "../actions"
+import {expandMenu} from "../actions/actions"
 
 export class NavBarMobile extends Component {
   onClick() {
@@ -24,4 +24,8 @@ render() {
 }
 }
 
-export default connect()(NavBarMobile);
+export const mapStateToProps = state => ({
+  drawerOpen: state.chart.drawerOpen
+})
+
+export default connect(mapStateToProps)(NavBarMobile);
