+++ src/enzymeAdapter.d.ts
@@ -1,0 +1,1 @@
+declare module "enzyme-adapter-react-16";
